refactor: extract visionproxy into server module.
